Change Migration Runner to use relative file path
Change the Migration Runner to identify files by their relative path
to the migration folder, instead of their filename. This allows files
with the same name to exist in different subdirectories of a migration
folder. Before this change RoundhousE only considered the filename when
checking if it had run the file previously.

e.g. Having a directory layout like this:
up
 |- feature1
 |   |- 001.sql
 |
 |- feature2
     |- 001.sql

Will now run successfully by roundhouse, previously it would identify
'feature2\001.sql' as being a modified copy of 'feature1\001.sql'.

Log messages and the script_name column in the ScriptsRun and
ScriptsRunErrors now store the script_name like 'feature1\001.sql'
instead of '001.sql'.

WARNING:
If you run this on a database already managed by RoundhousE that has
scripts inside subfolders of the migration folders it will re-run those
scripts, because only the filename portion was stored by versions of
RoundHouse before this commit.