Add option to run aftercreate with/out transaction
Add the runAfterCreateWithTransaction option which allows you to
independently control whether the after create scripts are run inside a
transaction or not. By default it is the same value as the
withTransaction option.

The rationale is that not all aftercreate scripts are transaction
compatiable and that since they are only run just after database
creation it should be fine to run outside a transaction.