Refactor searchallinsteadoftraverse handling
Add the FileLocator interface to abstract the selection of files from
the migration folders. This replaces the need for the Migration Runner
to explicitly decide what method to call on the FileSystemAccess
interface.

 - the Traverse class replaces the get_all_file_name_strings_in method.

 - the Recurse class replace the get_all_file_name_strings_recurevly_in
   method.

The dependency injection container picks the appropiate implementation
dependening on the prescene or lack of the /searchallinsteadoftraverse
switch.