Extract StatementSplitter interface
Extract StatementSplitter interface from the StatementSplitter class,
change the StatementSplitter class to be DefaultStatementSplitter and to
implement this interface.

Change the StatementSplitterSpecs so that they test the splitting of the
sql, not just the regex replaces the batch seperators with a replacement
string. Some modification was needed to handle whitespace added during
the split of a sql string.

Database interface has a property which retrieves the statement
splitter to use for that database type. This allows for different types
of databases to have custom Splitter implementation which handle the
specifics for that database. The default Database uses the
DefaultSplitter. The sql_statement_seperator_regex_pattern still remains
on the interface and DefaultDatabase class, but can probably be removed.