Add tokenizer based SqlServerStatementSplitter
Add implementation of StatementSplitter for SQL Server which tokenizers
and parses the sql to split it into batches. Fix bugs where statements
are split incorrectly between combinations of string literals and comments.

Regular expressions used to tokenize the sql statement, which is then
parsed to ensure that it is only split when outside comments and string
literals.

Remove whitespace from StatementSplitterSpecs because spurious
whitespace is not inserted by SqlServerStatementSplitter, unlike
DefaultStatementSplitter.