Add /storescriptrelativepath to toggle behaviour
Add StoreScriptRelativePath configuration option so that behaviour can
be toggled. The default is to have it off, so that just the filename is
used to determine if the script has been run previously. When switched
on the file's relative path to the migration directory is used to check
if it has been run before.