(GH-234) Fault tolerance for transient exceptions
Added transient exception handling for SqlServer database. This is to
have a better support for Azure SQL Database where transient faults
are expected and happen periodically.

Support for transient fault handling is added in a base DefaultDatabase
class, but the retry strategy implementation is specified in a
concrete database provider (mssql in this case).

Transient faults are handled in following places:
* Transient faults when opening a connection
* Transient faults when running an update script
* Transient faults when dealing with RoundhousE internal tables (via
NHibernate).

Microsoft TransientFaultHandling library was referenced for retry
building blocks. This might look like a dependency that could be
avoided (by just re-implementing necessary bits). It was mainly used to
utilize ReliableSqlConnection class.

As transient faults are much more common for Azure Sql Database than
for MS SQL Server, it might make sense to put retry logic into a
new dedicated Azure specific provider. But having it in a single MSSQL
provider is much more convenient from user perspective (less options to
think about) and it should not worsen experience for existing MSSQL
users.