(GH-237) Increase timeout for NHibernate commands
This is to prevent Timeout Expired exceptions when migration is running
on a slow/high latency server (the case for Azure Sql Server) or a
database with a lot of update scripts recorded in a `ScriptsRun` table.

The timeout of 5 minutes is currently hard-coded. Probably the better
way would be to add an additional configuration option. On the other
hand I'm not sure user should care about that additional config option.