Rework Line Ending Normalization => Hash Rules
The existing implementation did not account for mac line endings and did not normalize new hashes to be consistent.

This change does three things:
 - All future hashes are computed with line ending normalized to unix
 - Compares hashes with the unnormalized and the unix normalized (accounts for old hashes and new)
 - Compares the hash value with the line endings normalized to windows, unix, or mac

Any existing hash in the database that was calculated from a file with any consistent line ending should be pronounced equal. The only
case that we are not going to handle is if the existing hash was calculated from a file with mixed line endings and the line endings have
changed.