Change NuGet.exe reference in build config
The Nuget.exe in lib was removed in a prior commit (while the one in .nuget was updated). Without the lib version we
can't actually build packages for publishing. Rather than restore the redundant copy of the exe, I updated the build config
to reference the one that we do have.