Improve the implementation of the EnvironmentNames option
The original implementation passed a string for the multiple options and split the string later in the execution. This would put
library users in the unusual situation of having to specify a list of strings by concatenating a comma separated string. It also
preserved too much of the legacy EnvironmentName (singular) implementation.

This version:
- Changes Environment Names to an IList<string> that can be added, cleared, etc.
- Changes the implementation of the obsolete EnvironmentName property to use the Names list
- Splits the option string immediately